# Conflicts:
#	.gitmodules
#	Brewfile
#	CMakeLists.txt
#	README-Building-deps.txt
#	archive/windows/covise.iss
#	bin/hlrsDemo
#	build-vcpkg.bat
#	cmake/CoviseGlobalFeatureTests.cmake
#	cmake/CovisePresets.cmake
#	cmake/FindFFTW.cmake
#	cmake/Using/UseBoost.cmake
#	cmake/Using/UseVisionaray.cmake
#	config/config.hlrs.xml
#	config/config.xml
#	src/OpenCOVER/DrivingSim/TrafficSimulation/coEntity.cpp
#	src/OpenCOVER/DrivingSim/fasi2/fasi2.cpp
#	src/OpenCOVER/DrivingSim/mopla/Makefile
#	src/OpenCOVER/PluginUtil/ModuleFeedbackManager.cpp
#	src/OpenCOVER/cover/CMakeLists.txt
#	src/OpenCOVER/cover/OpenCOVER.cpp
#	src/OpenCOVER/cover/OpenCOVER.h
#	src/OpenCOVER/cover/VRRegisterSceneGraph.cpp
#	src/OpenCOVER/cover/VRWindow.cpp
#	src/OpenCOVER/cover/VruiPartnerMenuItem.cpp
#	src/OpenCOVER/cover/VruiPartnerMenuItem.h
#	src/OpenCOVER/cover/coVRAnimationManager.cpp
#	src/OpenCOVER/cover/coVRCollaboration.cpp
#	src/OpenCOVER/cover/coVRCollaboration.h
#	src/OpenCOVER/cover/coVRCommunication.cpp
#	src/OpenCOVER/cover/coVRCommunication.h
#	src/OpenCOVER/cover/coVRFileManager.cpp
#	src/OpenCOVER/cover/coVRFileManager.h
#	src/OpenCOVER/cover/coVRIntersectionInteractor.h
#	src/OpenCOVER/cover/coVRMSController.cpp
#	src/OpenCOVER/cover/coVRMSController.h
#	src/OpenCOVER/cover/coVRPluginSupport.cpp
#	src/OpenCOVER/cover/coVRPluginSupport.h
#	src/OpenCOVER/cover/input/person.cpp
#	src/OpenCOVER/cover/ui/Group.h
#	src/OpenCOVER/cover/ui/Manager.cpp
#	src/OpenCOVER/cover/ui/View.cpp
#	src/OpenCOVER/cover/ui/View.h
#	src/OpenCOVER/cover/ui/VruiView.cpp
#	src/OpenCOVER/cover/ui/VruiView.h
#	src/OpenCOVER/osgcaveui/Interaction.cpp
#	src/OpenCOVER/plugins/covise/COVISE/CovisePlugin.cpp
#	src/OpenCOVER/plugins/drivingsim/SumoTraCI/SumoTraCI.cpp
#	src/OpenCOVER/plugins/drivingsim/SumoTraCI/SumoTraCI.h
#	src/OpenCOVER/plugins/drivingsim/SumoTraCI/utils/traci/TraCIAPI.cpp
#	src/OpenCOVER/plugins/drivingsim/SumoTraCI/utils/traci/TraCIAPI.h
#	src/OpenCOVER/plugins/general/CMakeLists.txt
#	src/OpenCOVER/plugins/general/ColorBars/ColorBarPlugin.cpp
#	src/OpenCOVER/plugins/general/ColorBars/ColorBarPlugin.h
#	src/OpenCOVER/plugins/general/PointCloud/PointCloud.cpp
#	src/OpenCOVER/plugins/general/PointCloud/PointCloudInteractor.cpp
#	src/OpenCOVER/plugins/general/ViewPoint/ViewPoint.cpp
#	src/OpenCOVER/plugins/general/Vrml97/ExampleMatrixLight/PhotometricLightTest.wrl
#	src/OpenCOVER/plugins/general/Vrml97/ExampleMatrixLight/make_MatrixLightBinary.py
#	src/OpenCOVER/plugins/general/Vrml97/SystemCover.cpp
#	src/OpenCOVER/plugins/general/Vrml97/Vrml97Plugin.cpp
#	src/OpenCOVER/plugins/general/WindowTypeQt/WindowTypeQt.cpp
#	src/OpenCOVER/plugins/hlrs/CMakeLists.txt
#	src/OpenCOVER/plugins/hlrs/Midi/midiPlugin.cpp
#	src/OpenCOVER/plugins/hlrs/Midi/midiPlugin.h
#	src/OpenCOVER/plugins/hlrs/Revit/RevitPlugin/OpenCOVERPlugin/COVER.cs
#	src/OpenCOVER/plugins/hlrs/Revit/RevitPlugin/OpenCOVERPlugin/MessageBuffer.cs
#	src/OpenCOVER/plugins/hlrs/Spray/Spray.cpp
#	src/OpenCOVER/plugins/markertracking/common/RemoteAR.cpp
#	src/OpenCOVER/plugins/visenso/CfdGui/CfdGuiPlugin.cpp
#	src/OpenCOVER/plugins/visenso/DNA/DNABaseUnit.cpp
#	src/OpenCOVER/plugins/visenso/DocumentViewer/DocumentViewer.cpp
#	src/kernel/appl/ApplInterface.cpp
#	src/kernel/config/kernel/coConfigConstants.cpp
#	src/kernel/net/covise_socket.h
#	src/kernel/net/message.cpp
#	src/kernel/net/message_types.h
#	src/kernel/net/tokenbuffer.cpp
#	src/kernel/net/tokenbuffer.h
#	src/kernel/qtutil/FileSysAccess.cpp
#	src/kernel/qtutil/FileSysAccess.h
#	src/kernel/qtutil/NetHelp.cpp
#	src/kernel/qtutil/NetHelp.h
#	src/kernel/vrbclient/CMakeLists.txt
#	src/kernel/vrbclient/VRBClient.cpp
#	src/kernel/vrbclient/VRBClient.h
#	src/module/general/ReadFoam/archivemodel.cpp
#	src/sys/controller/CTRLHandler.h
#	src/sys/controller/Token.cpp
#	src/sys/controller/Token.h
#	src/sys/controller/coHostType.h
#	src/sys/vrb/VRBServer.cpp
#	src/sys/vrb/VRBServer.h
#	src/sys/vrb/gui/VRB.cpp
#	winenv.bat

+++ src/OpenCOVER/vrml97/vrml/Doc.cpp
@@ -71,7 +71,10 @@ Doc::~Doc()
     delete d_ostream;
     if (d_tmpfile)
     {
-        System::the->removeFile(d_tmpfile);
+		if (d_isTmp)
+		{
+			System::the->removeFile(d_tmpfile);
+		}
         delete[] d_tmpfile;
         d_tmpfile = 0;
     }
@@ -279,7 +282,7 @@ bool Doc::filename(char *fn, int nfn)
     fn[0] = '\0';
 
     char *e = 0, *s = (char *)stripProtocol(d_url);
-
+	std::string path;
     char *endString = NULL;
     int endLength = 0;
     if ((e = strrchr(s, '#')) != 0)
@@ -320,10 +323,11 @@ bool Doc::filename(char *fn, int nfn)
         // if not, the try to fetch it from a remote site if possible
         cerr << "file " << s << " not local, try to get it from remote " << endl;
 
-        if ((s = (char *)System::the->remoteFetch(d_url)))
+		path = System::the->remoteFetch(d_url, d_isTmp);
+		if (path != "")
         {
-            d_tmpfile = new char[strlen(s) + 1 + endLength];
-            strcpy(d_tmpfile, s);
+            d_tmpfile = new char[path.length() + 1 + endLength];
+            strcpy(d_tmpfile, path.c_str());
             //delete [] s; This causes a memory leak
             s = NULL;
             // XXX: there is a problem here: freeing sth not malloced...
@@ -341,28 +345,15 @@ bool Doc::filename(char *fn, int nfn)
     {
         // we have a local file now but does this exist?
         // if not, the try to fetch it from a remote site if possible
+		bool statbool;
 #ifdef WIN32
         struct _stat64 sbuf;
-        if (_stat64(s, &sbuf))
-        {
-            cerr << "file " << s << " not local, try to get it from remote " << endl;
-
-            if (d_tmpfile) // Already fetched it
-            {
-                s = d_tmpfile;
-            }
-            else if ((s = (char *)System::the->remoteFetch(s)))
-            {
-                d_tmpfile = new char[strlen(s) + 1 + endLength];
-                strcpy(d_tmpfile, s);
-                // XXX: there is a problem here: freeing sth not malloced...
-                //free(const_cast<char *>(s));        // assumes tempnam or equiv...
-                s = d_tmpfile;
-            }
-        }
+		statbool = _stat64(s, &sbuf);
 #else
-        struct stat sbuf;
-        if (stat(s, &sbuf))
+		struct stat sbuf;
+		statbool = stat(s, &sbuf);
+#endif
+        if (statbool)
         {
             cerr << "file " << s << " not local, try to get it from remote " << endl;
 
@@ -370,16 +361,16 @@ bool Doc::filename(char *fn, int nfn)
             {
                 s = d_tmpfile;
             }
-            else if ((s = (char *)System::the->remoteFetch(s)))
+            else if ((path = System::the->remoteFetch(s, d_isTmp)) != "")
             {
-                d_tmpfile = new char[strlen(s) + 1 + endLength];
-                strcpy(d_tmpfile, s);
+                d_tmpfile = new char[path.length() + 1 + endLength];
+                strcpy(d_tmpfile, path.c_str());
                 // XXX: there is a problem here: freeing sth not malloced...
                 //free(const_cast<char *>(s));        // assumes tempnam or equiv...
                 s = d_tmpfile;
             }
         }
-#endif
+
         if (s)
         {
             strncpy(fn, s, nfn - 1);
@@ -431,7 +422,10 @@ void Doc::fclose()
     d_fp = 0;
     if (d_tmpfile)
     {
-        System::the->removeFile(d_tmpfile);
+		if (d_isTmp)
+		{
+			System::the->removeFile(d_tmpfile);
+		}
         delete[] d_tmpfile;
         d_tmpfile = 0;
     }
@@ -482,7 +476,10 @@ void Doc::gzclose()
     d_gz = 0;
     if (d_tmpfile)
     {
-        System::the->removeFile(d_tmpfile);
+		if (d_isTmp)
+		{
+			System::the->removeFile(d_tmpfile);
+		}
         delete[] d_tmpfile;
         d_tmpfile = 0;
     }
